+++ Sources/SwiftDisc/Models/Guild.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct Guild: Codable, Hashable {
+    public let id: Snowflake
+    public let name: String
+}